+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## next (0.21.2)
+  - New top-level field: `extra-doc-files`.
+
 ## Changes in 0.21.1
   - Allow dependency constraints to be numbers (see #234)
 
@@ -5,7 +8,6 @@
   - Accept section-specific fields in conditionals (see #175, thanks to Scott
     Fleischman)
   - New section: `internal-libraries`, for Cabal 2's internal libraries (see #200).
-  - New top-level field: `extra-doc-files`.
 
 ## Changes in 0.20.0
   - Do not overwrite any existing cabal file if it has been modified manually